Rip out queryparse tests away from test_dbcore
